commit title : create home screen ui and sheet component
what : create searchBar and gridView of sheet

why : sheet component is always to use every page So I think it should be built first. Create home page so that everyone can see the image of the app.

how : Use SingleChildScrollView to be able to scroll down and use gridView to make our app responsive.