Commit title: Fix login screen again kub
What: Fix login screen and rename the mis-spelling function

Why: Because before fixing. When you open profile page, and then login, it pass, and then logout, and then login again, it will be the circularprogressindicator like not login. It's happen because the streambuilder is not building new screen that their condition is when data is changing, but when login/logout is not changing any of data. So it's not update screen

How: Just do like sidebar screen. Use streambuilder with authStateChanged to build new screen when login/logout and add rest of the code inside true contition