Commit title: Google sign-in system
What: Add method to sign-in with google in login and register page. And add method to create doc in firestore after first time login

Why: Because it's easier to login compare to email/password

How: It's have much process before the code. But code is very similar with email password login method.