Commit title: Fetch user while logon
What: Currently, if some fields have been created and user not have it. It will create that field automatically

Why: Because it's very annoying when add new data to model and need to clean old data because it's not valid

How: Use containsKey to check field. If it's have, it will contain same data. But if it's not have it'll create new field. However, DON'T DELETE some field such as username, email, uid and profileImage because current condition it will be null and it will cause bug