Commit title: Purchase coin with credit/debit card
What: Success creating method that will add coin to user via credit/debit card. But cannot paid with added card because it's not finished

Why: To make app more complete

How: I watch same Indian guy as mvvm guy. To tell brief concept, it use createPaymentIntent to create paid screen that will popup from bottom and it can setup some data to paid. For displayPaymentSheet is the method that handle if it succeed or not. And calculateAmount is same with it's name