Commit title: Buysheet flushbar fix and add more condition
What: Fix flushbar when buying sheet and add condition to skip this method when price is 0 and cause other glitch

Why: Because user will know easier when sheet is buyed, and for condition. It's will speed-up open sheet time a lot because before the fix it do everything that it shouldn't do

How: Move code to move above part of method and add condition to check price of sheet, if price is zero. Bye bye, it works