Commit title: Cover image fix
What: Mainly of this commit is add method to fix the imageToFile method. And a little fix is change flushbar icon when sheet creation is finished.

Why: This commit I hopefully it's will not make cover image that bugged again. And for flushbar icon it is just mis-using icon.

How: Before I start to apply this method, I check the broked file when bug is occur. It's 0 bytes image file. So, I counted bytes of image first. Then, If filesize is zero, i'll get into while loop to loop converting file. That's it.