Commit title: Fix in image_api files too and some other bugs
What: Fix in image_api that use same function like in pdf_api files, and fix some sort of bugs in other files such as like and dislike of question is null and question pages that check on null values.

Why: To make sure that imageToFile function will works finely and other page will work fine without bugs too.

How: I change from apply logic inside imageToFile method to be create new method for it to handle image because it's prettier. The two new function are doing same thing, but the reason that I not use the same function is in a writeAsBytes it's use different input, so I need to create method for it. And other bugs it just checking wrong snapshot, so I'm fix it. And for zero like bug, I assign value into it at the declaration.